Move disparate helpers to new core.lib module.
