+++ core/lib/text.py
@@ -1,0 +1,21 @@
+import random
+import string
+import unicodedata
+
+
+def contains_digits(string):
+    return any(char.isdigit() for char in string)
+
+
+def random_string(self, length):
+    return "".join(random.choice(string.ascii_letters) for i in range(length))
+
+
+def remove_accents(input_str):
+    # see https://stackoverflow.com/a/517974/330911
+    nfkd_form = unicodedata.normalize("NFKD", input_str)
+    return "".join([c for c in nfkd_form if not unicodedata.combining(c)])
+
+
+def ucfirst(string):
+    return string[0].upper() + string[1:]
